Make it compilable under modern Linuxes

+++ intrace.h
@@ -30,7 +30,7 @@ typedef struct {
 
 	char *hostname;
 	uint16_t port;
-	unsigned int paylSz;
+	size_t paylSz;
 
 	struct in_addr rip;
 	struct in_addr lip;
